requirements.txt
pip
croniter
pywaggle[vision] == 0.57.*